requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
torch>=2.0.0
transformers>=4.30.0
torchvision>=0.15.0
Pillow>=9.0.0
scikit-learn>=1.0.0